[PT 6] Filtering and order history

+++ data/lib/entity/shopping_cart_item/shopping_cart_item_entity_adapter.dart
@@ -9,7 +9,7 @@ class ShoppingCartItemEntityAdapter
   ShoppingCartItemEntity read(BinaryReader reader) {
     return ShoppingCartItemEntity(
       menuItemEntity: reader.read(),
-      amount: reader.read() as int,
+      amount: reader.read(),
     );
   }
